ioremap: Adding functionality to support ioremap
ioremap maps given physical address to the current page table.
This allows drivers to work with virtual memory while retaining
their functionality.

NOTE: Currently all entries only have default/generic flags, so
memory operations with peripherals may not be reliable

NOTE: this currently does not work AFTER switching execution
to high memory (needs further investigation)

+++ hobos/lib/pt_lib.h
@@ -83,8 +83,8 @@ struct page_table_desc {
 };
 
 struct va_metadata {
-	uint32_t index[3];
-	uint16_t offset;
+	uint64_t index[3];
+	uint64_t offset;
 };
 
 //extract pte desc/meta data - more for debugging purposes
